requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
python-multipart==0.0.6
pandas==2.1.1
python-dotenv==1.0.0
google-generativeai==0.3.1

